기능 추가 : 방 조회 시 방명록 조회 추가
