+++ src/main/java/com/sorhive/comprojectserver/file/S3LifingImageFile.java
@@ -1,31 +1,30 @@
-package com.sorhive.comprojectserver.config.file;
+package com.sorhive.comprojectserver.file;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
-import net.coobird.thumbnailator.Thumbnails;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayInputStream;
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Optional;
 
 /**
  * <pre>
- * Class : S3File
+ * Class : S3LifingImageFile
  * Comment: 클래스에 대한 간단 설명
  * History
  * ================================================================
  * DATE             AUTHOR           NOTE
  * ----------------------------------------------------------------
- * 2022-10-24       부시연           최초 생성
+ * 2022-11-11      부시연           최초 생성
  * </pre>
  *
  * @author 부시연(최초 작성자)
@@ -33,31 +32,38 @@ import java.util.Optional;
  */
 @Component
 @RequiredArgsConstructor
-public class S3File {
-    private static final Logger log = LoggerFactory.getLogger(S3File.class);
+public class S3LifingImageFile {
+    private static final Logger log = LoggerFactory.getLogger(S3LifingImageFile.class);
     private final AmazonS3Client amazonS3Client;
 
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-    public String upload(MultipartFile multipartFile , String dirName) throws IOException {
+    public String upload(byte[] lifingImage, String changeName, String dirName) throws IOException {
 
-        File uploadFile = convert(multipartFile)
-                .orElseThrow(() -> new IllegalArgumentException("MultipartFile -> File로 전환이 실패했습니다."));
+        ByteArrayInputStream input_stream= new ByteArrayInputStream(lifingImage);
 
-        return upload(uploadFile, dirName);
+        BufferedImage final_buffered_image = ImageIO.read(input_stream);
+
+        ImageIO.write(final_buffered_image , "png", new File("./" + changeName) );
+
+        File newFile = new File("./" + changeName);
+
+        return upload(newFile, changeName, dirName);
     }
 
-    private String upload(File uploadFile, String dirName) throws IOException {
+    private String upload(File uploadFile, String changeName, String dirName) {
 
-        String fileName = dirName + "/" + uploadFile.getName();
+        String fileName = dirName + "/" + changeName;
         String uploadImageUrl = putS3(uploadFile, fileName);
         removeNewFile(uploadFile);
         return uploadImageUrl;
     }
 
     private String putS3(File uploadFile, String fileName) {
-        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile).withCannedAcl(CannedAccessControlList.PublicRead));
+        amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+
         return amazonS3Client.getUrl(bucket, fileName).toString();
     }
 
@@ -67,18 +73,6 @@ public class S3File {
         } else {
             log.info("파일이 삭제되지 못했습니다.");
         }
-    }
-
-    private Optional<File> convert(MultipartFile file) throws IOException {
-        File convertFile = new File(file.getOriginalFilename());
-        if(convertFile.createNewFile()) {
-            try (FileOutputStream fos = new FileOutputStream(convertFile)) {
-                fos.write(file.getBytes());
-            }
-            return Optional.of(convertFile);
-        }
-
-        return Optional.empty();
     }
 
     public void delete(String fileName) {
