+++ src/main/java/com/sorhive/comprojectserver/file/S3MemberFile.java
@@ -1,8 +1,7 @@
-package com.sorhive.comprojectserver.config.file;
+package com.sorhive.comprojectserver.file;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -12,19 +11,18 @@ import org.springframework.stereotype.Component;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
-import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
+import java.io.*;
 
 /**
  * <pre>
- * Class : S3LifingImageFile
+ * Class : S3MemberRoomFile
  * Comment: 클래스에 대한 간단 설명
  * History
  * ================================================================
  * DATE             AUTHOR           NOTE
  * ----------------------------------------------------------------
- * 2022-11-11      부시연           최초 생성
+ * 2022-11-10       부시연           최초 생성
+ * 2022-11-11       부시연           바이트 배열 대응 추가
  * </pre>
  *
  * @author 부시연(최초 작성자)
@@ -32,35 +30,34 @@ import java.io.IOException;
  */
 @Component
 @RequiredArgsConstructor
-public class S3LifingImageFile {
-    private static final Logger log = LoggerFactory.getLogger(S3LifingImageFile.class);
+public class S3MemberFile {
+    private static final Logger log = LoggerFactory.getLogger(S3MemberFile.class);
     private final AmazonS3Client amazonS3Client;
 
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
-    public String upload(byte[] lifingImage, String changeName, String dirName) throws IOException {
+    public String upload(byte[] roomFile , String dirName, String member) throws IOException {
 
-        ByteArrayInputStream input_stream= new ByteArrayInputStream(lifingImage);
-
+        ByteArrayInputStream input_stream= new ByteArrayInputStream(roomFile);
         BufferedImage final_buffered_image = ImageIO.read(input_stream);
+        ImageIO.write(final_buffered_image , "png", new File("./temp.png") );
+        File newFile = new File("./temp.png");
 
-        ImageIO.write(final_buffered_image , "png", new File("./" + changeName) );
-
-        File newFile = new File("./" + changeName);
-
-        return upload(newFile, changeName, dirName);
+        return upload(newFile, dirName, member);
     }
 
-    private String upload(File uploadFile, String changeName, String dirName) {
+    private String upload(File uploadFile, String dirName, String member) {
 
-        String fileName = dirName + "/" + changeName;
+        String fileName = dirName + "/" + member;
         String uploadImageUrl = putS3(uploadFile, fileName);
         removeNewFile(uploadFile);
+
         return uploadImageUrl;
     }
 
     private String putS3(File uploadFile, String fileName) {
+
         amazonS3Client.putObject(new PutObjectRequest(bucket, fileName, uploadFile)
                 .withCannedAcl(CannedAccessControlList.PublicRead));
 
@@ -68,15 +65,13 @@ public class S3LifingImageFile {
     }
 
     private void removeNewFile(File targetFile) {
+
         if (targetFile.delete()) {
             log.info("파일이 삭제되었습니다.");
         } else {
             log.info("파일이 삭제되지 못했습니다.");
         }
+
     }
 
-    public void delete(String fileName) {
-        DeleteObjectRequest request = new DeleteObjectRequest(bucket, "images/" + fileName);
-        amazonS3Client.deleteObject(request);
-    }
 }
