+++ src/main/java/com/sorhive/comprojectserver/lifing/command/infra/LifingService.java
@@ -1,6 +1,6 @@
 package com.sorhive.comprojectserver.lifing.command.infra;
 
-import com.sorhive.comprojectserver.config.file.S3LifingImageFile;
+import com.sorhive.comprojectserver.file.S3LifingImageFile;
 import com.sorhive.comprojectserver.config.jwt.TokenProvider;
 import com.sorhive.comprojectserver.lifing.command.application.dto.LifingCreateDto;
 import com.sorhive.comprojectserver.lifing.command.application.dto.LifingImageDto;
